+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/TransitRouter.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.routing.algorithm.raptor.router;
 
-import java.time.Instant;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -47,23 +47,26 @@ public class TransitRouter {
     private final RoutingRequest request;
     private final Router router;
     private final DebugTimingAggregator debugTimingAggregator;
+    private final ZonedDateTime searchStartTime;
 
     private TransitRouter(
             RoutingRequest request,
             Router router,
+            ZonedDateTime searchStartTime,
             DebugTimingAggregator debugTimingAggregator
     ) {
         this.request = request;
         this.router = router;
         this.debugTimingAggregator = debugTimingAggregator;
+        this.searchStartTime = searchStartTime;
     }
 
     private TransitRouterResult route() {
         if (request.modes.transitModes.isEmpty()) {
-            return new TransitRouterResult(List.of(), null, NOT_SET);
+            return new TransitRouterResult(List.of(), null);
         }
 
-        if (!router.graph.transitFeedCovers(request.dateTime)) {
+        if (!router.graph.transitFeedCovers(request.getDateTimeOriginalSearch())) {
             throw new RoutingValidationException(List.of(
                     new RoutingError(RoutingErrorCode.OUTSIDE_SERVICE_PERIOD, InputField.DATE_TIME)
             ));
@@ -91,7 +94,7 @@ public class TransitRouter {
         // Prepare transit search
         var raptorRequest = RaptorRequestMapper.mapRequest(
                 request,
-                requestTransitDataProvider.getStartOfTime(),
+                searchStartTime,
                 accessEgresses.getAccesses(),
                 accessEgresses.getEgresses()
         );
@@ -123,7 +126,7 @@ public class TransitRouter {
         RaptorPathToItineraryMapper itineraryMapper = new RaptorPathToItineraryMapper(
                 router.graph,
                 transitLayer,
-                requestTransitDataProvider.getStartOfTime(),
+                searchStartTime,
                 request
         );
         FareService fareService = router.graph.getService(FareService.class);
@@ -140,19 +143,9 @@ public class TransitRouter {
             itineraries.add(itinerary);
         }
 
-        // Filter itineraries away that depart after the latest-departure-time for depart after
-        // search. These itineraries is a result of time-shifting the access leg and is needed for
-        // the raptor to prune the results. These itineraries are often not ideal, but if they
-        // pareto optimal for the "next" window, they will appear when a "next" search is performed.
-        Instant filterOnLatestDepartureTime = null;
-        var searchWindowUsedInSeconds = transitResponse.requestUsed().searchParams().searchWindowInSeconds();
-        if(!request.arriveBy && searchWindowUsedInSeconds > 0) {
-            filterOnLatestDepartureTime = Instant.ofEpochSecond(request.dateTime + searchWindowUsedInSeconds);
-        }
-
         debugTimingAggregator.finishedItineraryCreation();
 
-        return new TransitRouterResult(itineraries, filterOnLatestDepartureTime, searchWindowUsedInSeconds);
+        return new TransitRouterResult(itineraries, transitResponse.requestUsed().searchParams());
     }
 
     private AccessEgresses getAccessEgresses(
@@ -248,7 +241,7 @@ public class TransitRouter {
             return new RaptorRoutingRequestTransitData(
                     graph.getTransferService(),
                     transitLayer,
-                    request.getDateTime().toInstant(),
+                    searchStartTime,
                     request.arriveBy ? request.additionalSearchDaysBeforeToday : 0,
                     request.arriveBy ? 0 : request.additionalSearchDaysAfterToday,
                     createRequestTransitDataProviderFilter(graph.index),
@@ -298,8 +291,9 @@ public class TransitRouter {
     public static TransitRouterResult route(
             RoutingRequest request,
             Router router,
+            ZonedDateTime searchStartTime,
             DebugTimingAggregator debugTimingAggregator
     ) {
-        return new TransitRouter(request, router, debugTimingAggregator).route();
+        return new TransitRouter(request, router, searchStartTime, debugTimingAggregator).route();
     }
 }