+++ src/main/java/org/opentripplanner/routing/api/request/RoutingRequest.java
@@ -4,6 +4,7 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.Serializable;
 import java.time.Duration;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -34,6 +35,7 @@ import org.opentripplanner.model.Route;
 import org.opentripplanner.model.TransitMode;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilter;
 import org.opentripplanner.routing.algorithm.transferoptimization.api.TransferOptimizationParameters;
+import org.opentripplanner.model.plan.PageCursor;
 import org.opentripplanner.routing.core.BicycleOptimizeType;
 import org.opentripplanner.routing.core.RouteMatcher;
 import org.opentripplanner.routing.core.RoutingContext;
@@ -52,6 +54,7 @@ import org.opentripplanner.routing.spt.ShortestPathTree;
 import org.opentripplanner.routing.vehicle_rental.RentalVehicleType.FormFactor;
 import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
 import org.opentripplanner.util.time.DateUtils;
+import org.opentripplanner.util.time.DurationUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -78,6 +81,8 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
     private static final long serialVersionUID = 1L;
 
     private static final Logger LOG = LoggerFactory.getLogger(RoutingRequest.class);
+
+    private static final long NOW_THRESHOLD_SEC = DurationUtils.duration("15h");
 
     /* FIELDS UNIQUELY IDENTIFYING AN SPT REQUEST */
 
@@ -159,8 +164,11 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
      */
     public BicycleOptimizeType bicycleOptimizeType = BicycleOptimizeType.SAFE;
 
-    /** The epoch date/time that the trip should depart (or arrive, for requests where arriveBy is true) */
-    public long dateTime = new Date().getTime() / 1000;
+    /**
+     * The epoch date/time in seconds that the trip should depart (or arrive, for requests where
+     * arriveBy is true)
+     */
+    private long dateTime = new Date().getTime() / 1000;
 
     /**
      * This is the time/duration in seconds from the earliest-departure-time(EDT) to
@@ -173,13 +181,27 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
      * Transit search as well; Hence this is named search-window and not raptor-search-window. Do
      * not confuse this with the travel-window, which is the time between EDT to LAT.
      * <p>
-     * Use {@code null} to unset, and {@link Duration#ZERO} to do one Raptor iteration. The value is
-     * dynamically  assigned a suitable value, if not set. In a small to medium size operation
+     * Use {@code null} to unset, and {@link Duration#ZERO} to do one Raptor iteration. The value
+     * is dynamically  assigned a suitable value, if not set. In a small to medium size operation
      * you may use a fixed value, like 60 minutes. If you have a mixture of high frequency cities
      * routes and infrequent long distant journeys, the best option is normally to use the dynamic
      * auto assignment.
+     * <p>
+     * There is no need to set this when going to the next/previous page any more.
      */
     public Duration searchWindow;
+
+    /**
+     * Use the cursor to go to the next or previous "page" of trips.
+     * You should pass in the original request as is.
+     * <p>
+     * The next page of itineraries will depart after the current results
+     * and the previous page of itineraries will depart before the current results.
+     * <p>
+     * The paging does not support timeTableView=false and arriveBy=true, this will result in
+     * none pareto-optimal results.
+     */
+    public PageCursor pageCursor;
 
     /**
      * Search for the best trip options within a time window. If {@code true} two itineraries are
@@ -946,17 +968,41 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
         streetSubRequestModes.setMode(mode, true);
     }
 
-    public Date getDateTime() {
-        return new Date(dateTime * 1000);
+    /**
+     * When a client perform the first search it supply a search time - its is that
+     * time. The client may go to the next page, but the original datetime stay unchanged.
+     */
+    public Instant getDateTimeOriginalSearch() {
+        return Instant.ofEpochSecond(dateTime);
     }
 
-    public void setDateTime(Date dateTime) {
-        this.dateTime = dateTime.getTime() / 1000;
+    /**
+     * The search time for the current page. If the client have moved to the next page
+     * then this is the adjusted search time. The search time is adjusted with according to
+     * the time-window used.
+     */
+    public Instant getDateTimeCurrentPage() {
+        return pageCursor == null ? Instant.ofEpochSecond(dateTime) : (
+                arriveBy
+                        ? pageCursor.latestArrivalTime
+                        : pageCursor.earliestDepartureTime
+        );
+    }
+
+    public void setDateTime(Instant dateTime) {
+        this.dateTime = dateTime.getEpochSecond();
     }
 
     public void setDateTime(String date, String time, TimeZone tz) {
         Date dateObject = DateUtils.toDate(date, time, tz);
-        setDateTime(dateObject);
+        setDateTime(dateObject == null ? Instant.now() : dateObject.toInstant());
+    }
+
+    /**
+     * Is the trip originally planned withing the previous/next 15h?
+     */
+    public boolean isTripPlannedForNow() {
+        return Math.abs(Instant.now().getEpochSecond() - dateTime) < NOW_THRESHOLD_SEC;
     }
 
     /**
@@ -964,6 +1010,10 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
      */
     public int getNumItineraries() {
         return 1;
+    }
+
+    public void setPageCursor(String pageCursor) {
+        this.pageCursor = PageCursor.decode(pageCursor);
     }
 
     public void setNumItineraries(int numItineraries) {
@@ -975,7 +1025,7 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
     }
 
     public String toString(String sep) {
-        return from + sep + to + sep + getDateTime() + sep
+        return from + sep + to + sep + getDateTimeOriginalSearch() + sep
                 + arriveBy + sep + bicycleOptimizeType + sep + streetSubRequestModes.getAsStr() + sep
                 + getNumItineraries();
     }