+++ src/test/java/org/opentripplanner/routing/algorithm/TestAStar.java
@@ -1,20 +1,21 @@
 package org.opentripplanner.routing.algorithm;
 
+import static org.opentripplanner.gtfs.GtfsContextBuilder.contextBuilder;
+
+import java.time.Instant;
 import junit.framework.TestCase;
 import org.junit.Ignore;
 import org.opentripplanner.ConstantsForTests;
+import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.gtfs.GtfsContext;
 import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.api.request.RoutingRequest;
-import org.opentripplanner.graph_builder.module.geometry.GeometryAndBlockProcessor;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.spt.GraphPath;
 import org.opentripplanner.routing.spt.ShortestPathTree;
 import org.opentripplanner.util.TestUtils;
-
-import static org.opentripplanner.gtfs.GtfsContextBuilder.contextBuilder;
 
 /**
  * TODO OTP2 - Test is too close to the implementation and will need to be reimplemented.
@@ -40,7 +41,7 @@ public class TestAStar extends TestCase {
         GraphPath path = null;
 
         String feedId = gg.getFeedIds().iterator().next();
-        options.dateTime = TestUtils.dateInSeconds("America/Los_Angeles", 2009, 8, 7, 12, 0, 0);
+        options.setDateTime(TestUtils.dateInstant("America/Los_Angeles", 2009, 8, 7, 12, 0, 0));
         options.setRoutingContext(gg, feedId + ":Millbrae Caltrain", feedId + ":Mountain View Caltrain");
         spt = aStar.getShortestPathTree(options);
         path = spt.getPath(gg.getVertex(feedId + ":Mountain View Caltrain"), true);
@@ -51,7 +52,7 @@ public class TestAStar extends TestCase {
 
         /* test backwards traversal */
         options.setArriveBy(true);
-        options.dateTime = endTime;
+        options.setDateTime(Instant.ofEpochSecond(endTime));
         options.setRoutingContext(gg, feedId + ":Millbrae Caltrain", feedId + ":Mountain View Caltrain");
         spt = aStar.getShortestPathTree(options);
         path = spt.getPath(gg.getVertex(feedId + ":Millbrae Caltrain"), true);
@@ -71,7 +72,7 @@ public class TestAStar extends TestCase {
 
         RoutingRequest options = new RoutingRequest();
         long startTime = TestUtils.dateInSeconds("America/Los_Angeles", 2009, 11, 1, 12, 34, 25);
-        options.dateTime = startTime;
+        options.setDateTime(Instant.ofEpochSecond(startTime));
         // one hour is more than enough time
         options.setRoutingContext(graph, start, end);